refactor(app): extract scroll-sync orchestrators to scroll_sync (phase 3f)
Move ensure_active_tab_visible, sync_scroll_groups, and
pre_sync_ensure_visible out of render.rs into a new app/scroll_sync.rs
sibling. ~334 lines.

The latter two were `fn` (private to render.rs); they become
`pub(super)` so the in-render `recompute_layout` can still call them.

render.rs goes from 2916 -> 2582 lines (-334). 2026 tests still pass.
